# docs/openapi.json
{
  "info": {
    "description": "Personal finance tracker API.\n\nRequests that change data (POST, PUT, PATCH, DELETE) must send any body as `application/json` (415 otherwise). A browser may send them only from the same origin or an origin listed in `CORS_ORIGINS` (403 otherwise).",
    "title": "Vipu API",
    "version": "0.1.0"
  },
...
  "servers": [
    {
      "name": "Local",
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
...
    "/api/budget/snapshots/{snapshot_id}": {
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "tags": [
          "Budget Snapshots"
        ],
        "summary": "Update a budget snapshot's entries and/or notes.",
        "description": "Accepts JSON with:\n- ``entries``: list of ``{account_name, balance, is_credit, account_id?}``\n- ``notes``: optional string (max 500 chars)\n\nRecalculates current_balance and change_from_previous. Also updates\nthe next snapshot's change_from_previous if one exists."
      },
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "tags": [
          "Budget Snapshots"
        ],
        "summary": "Delete a budget snapshot.",
        "description": "After deletion, the next snapshot's ``change_from_previous`` is\nrecalculated against the new previous snapshot so the displayed\ndelta stays consistent."
      }
    },
...
  },
  "openapi": "3.1.0",
  "components": {
    "schemas": {
      "HTTPError": {
        "properties": {
          "detail": {
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      }
    }
  }
}